requirements.txt
abutils
click
magika
natsort
numpy
prody
pyyaml
torch
tqdm